requirements.txt
nicegui>=1.4.20
numpy>=1.26
Pillow>=10.0.0
